LPS-123183 Create a class that handles the locale changes so it can be reused
